moved tests to appropritate file
