+++ tests/utils/test_utils.py
@@ -4,6 +4,8 @@ import pytest
 import sys
 import copy
 import collections
+import logging
+
 from pathlib import Path
 
 from BALSAMIC.utils.cli import SnakeMake
@@ -29,6 +31,10 @@ from BALSAMIC.utils.rule import get_picard_mrkdup
 from BALSAMIC.utils.rule import get_script_path
 from BALSAMIC.utils.rule import get_result_dir
 from BALSAMIC.utils.rule import get_threads
+from BALSAMIC.utils.cli import (merge_json, validate_fastq_pattern,
+                                get_panel_chrom, get_bioinfo_tools_list,
+                                get_sample_dict, get_sample_names,
+                                create_fastq_symlink)
 
 
 def test_recursive_default_dict():
@@ -459,3 +465,83 @@ def test_find_file_index(tmpdir):
     assert isinstance(result, list)
     assert str(bai_file) in result
     assert str(bai_file_2) in result
+
+
+def test_merge_json(config_files):
+    # GIVEN a dict and json file
+    ref_dict = json.load(open(config_files['reference'], 'r'))
+
+    json_file = config_files['sample']
+
+    # WHEN passing dict and json file to merge
+    merge_dict = merge_json(ref_dict, json_file)
+
+    # THEN It will merge both the data and return dict
+    assert isinstance(merge_dict, dict)
+    assert "samples" in merge_dict
+    assert "reference" in merge_dict
+
+
+def test_merge_json_error(config_files):
+    with pytest.raises(Exception, match=r"No such file or directory"):
+        # GIVEN a dict and invalid json file path
+        ref_dict = json.load(open(config_files['reference'], 'r'))
+        json_file = 'reference.json'
+
+        # WHEN passing python dict and invalid json path
+        # THEN it should throw OSError as FileNotFoundError
+        assert merge_json(ref_dict, json_file)
+
+
+def test_validate_fastq_pattern():
+    #GIVEN a path to a file with correct fastq file prefix
+    fastq_path_r1 = "/home/analysis/dummy_tumor_R_1.fastq.gz"
+    fastq_path_r2 = "/home/analysis/dummy_normal_R_2.fastq.gz"
+    #THEN it should return the correct prefix
+    assert validate_fastq_pattern(fastq_path_r1) == "dummy_tumor_R"
+    assert validate_fastq_pattern(fastq_path_r2) == "dummy_normal_R"
+
+    with pytest.raises(AttributeError) as excinfo:
+        #GIVEN a path to a file with incorrect fastq file prefix    
+        bad_fastq_path_1 = "/home/analysis/dummy_tumor.fastq.gz"
+        validate_fastq_pattern(bad_fastq_path_1)
+        #THEN AttributeError is raised
+    assert excinfo.value
+
+    with pytest.raises(AttributeError) as excinfo:
+        #GIVEN a path to a file with incorrect fastq file prefix   
+        bad_fastq_path_2 = "/home/analysis/dummy_tumor_R3.fastq.gz"
+        validate_fastq_pattern(bad_fastq_path_2)
+        #THEN AttributeError is raised
+    assert excinfo.value
+
+    with pytest.raises(AttributeError) as excinfo:
+        #GIVEN a path to a file with incorrect fastq file prefix   
+        bad_fastq_path_3 = "/home/analysis/dummy_tumor_R_2.bam"
+        validate_fastq_pattern(bad_fastq_path_3)
+        #THEN AttributeError is raised
+    assert excinfo.value
+
+
+def test_get_panel_chrom():
+    #GIVEN a valid PANEL BED file
+    panel_bed_file = 'tests/test_data/references/panel/panel.bed'
+    #THEN it should return a set containing multiple unique chromosomes
+    assert len(get_panel_chrom(panel_bed_file)) > 0
+
+
+def test_create_fastq_symlink(tmpdir_factory, caplog):
+    #GIVEN a list of valid input fastq files from test directory containing 4 files
+    symlink_from_path = tmpdir_factory.mktemp("symlink_from")
+    symlink_to_path = tmpdir_factory.mktemp("symlink_to")
+    filenames = ["tumor_R_1.fastq.gz", "normal_R_1.fastq.gz", "tumor_R_2.fastq.gz", "normal_R_2.fastq.gz"]
+    successful_log = "skipping"
+    casefiles = [Path(symlink_from_path,x) for x in filenames]
+    for casefile in casefiles:
+        casefile.touch()
+    with caplog.at_level(logging.INFO):
+        create_fastq_symlink(casefiles=casefiles, symlink_dir=symlink_to_path)
+        #THEN destination should have 4 files
+        assert len(list(Path(symlink_to_path).rglob("*.fastq.gz"))) == 4
+        #THEN exception triggers log message containign "skipping"
+        assert successful_log in caplog.text
